Identity: update id type to Long.

+++ src/main/java/com/ibanfr/infrastructure/dao/InMemoryDao.java
@@ -1,6 +1,6 @@
 package com.ibanfr.infrastructure.dao;
 
-import com.ibanfr.domain.Identity;
+import com.ibanfr.domain.model.Identity;
 
 import java.util.*;
 
@@ -10,12 +10,12 @@ public class InMemoryDao<T extends Identity> implements Dao<T>{
 
     @Override
     public void save(T entity) {
-        entity.setId(new Random().nextInt());
+        entity.setId(new Random().nextLong());
         repository.add(entity);
     }
 
     @Override
-    public Optional<T> findById(Integer id) {
+    public Optional<T> findById(Long id) {
         return repository.stream()
                 .filter(t -> t.getId()
                         .equals(id))
